+++ src/compound.rs
@@ -4,12 +4,16 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use canonical::{Canon, CanonError};
+
 use core::marker::PhantomData;
 
-use crate::annotations::{Annotated, Annotation, WrappedAnnotation};
-use canonical::Canon;
+use crate::annotations::{Annotation, WrappedAnnotation};
+use crate::generic::GenericTree;
+use crate::link::Link;
 
 /// The response of the `child` method on a `Compound` node.
+#[derive(Debug)]
 pub enum Child<'a, C, A>
 where
     C: Compound<A>,
@@ -17,7 +21,7 @@ where
     /// Child is a leaf
     Leaf(&'a C::Leaf),
     /// Child is an annotated subtree node
-    Node(&'a Annotated<C, A>),
+    Node(&'a Link<C, A>),
     /// Empty slot
     Empty,
     /// No more children
@@ -25,6 +29,7 @@ where
 }
 
 /// The response of the `child_mut` method on a `Compound` node.
+#[derive(Debug)]
 pub enum ChildMut<'a, C, A>
 where
     C: Compound<A>,
@@ -32,7 +37,7 @@ where
     /// Child is a leaf
     Leaf(&'a mut C::Leaf),
     /// Child is an annotated node
-    Node(&'a mut Annotated<C, A>),
+    Node(&'a mut Link<C, A>),
     /// Empty slot
     Empty,
     /// No more children
@@ -40,68 +45,79 @@ where
 }
 
 /// A type that can recursively contain itself and leaves.
-pub trait Compound<A>: Sized + Canon {
+pub trait Compound<A>: Canon {
     /// The leaf type of the Compound collection
-    type Leaf;
+    type Leaf: Canon;
 
     /// Returns a reference to a possible child at specified offset
-    fn child(&self, ofs: usize) -> Child<Self, A>
-    where
-        A: Annotation<Self::Leaf>;
+    fn child(&self, ofs: usize) -> Child<Self, A>;
 
     /// Returns a mutable reference to a possible child at specified offset
-    fn child_mut(&mut self, ofs: usize) -> ChildMut<Self, A>
-    where
-        A: Annotation<Self::Leaf>;
+    fn child_mut(&mut self, ofs: usize) -> ChildMut<Self, A>;
 
-    /// Returns an iterator over the children of the Compound node.
-    fn children(&self) -> ChildIterator<Self, A> {
-        ChildIterator {
+    /// Provides an iterator over all sub-annotations of the compound node
+    fn annotations(&self) -> AnnoIter<Self, A>
+    where
+        A: Annotation<Self::Leaf>,
+    {
+        AnnoIter {
             node: self,
             ofs: 0,
             _marker: PhantomData,
         }
     }
-}
 
-/// The kinds of children you can encounter iterating over a Compound
-pub enum IterChild<'a, C, A>
-where
-    C: Compound<A>,
-{
-    /// Iterator found a leaf
-    Leaf(&'a C::Leaf),
-    /// Iterator found an annotated node
-    Node(&'a Annotated<C, A>),
-}
+    /// Returns a generic version of this compound tree, erasing the specific
+    /// annotation and leaf types, to provide a universal tree encoding.
+    fn generic(&self) -> GenericTree
+    where
+        Self::Leaf: Canon,
+        A: Annotation<Self::Leaf>,
+    {
+        let mut generic = GenericTree::new();
 
-impl<'a, C, A> IterChild<'a, C, A>
-where
-    A: Annotation<C::Leaf>,
-    C: Compound<A>,
-{
-    /// Returns the annotation of the child
-    pub fn annotation(&self) -> WrappedAnnotation<A> {
-        match self {
-            IterChild::Leaf(l) => WrappedAnnotation::Owned(A::from_leaf(l)),
-            IterChild::Node(a) => WrappedAnnotation::Borrowed(a.annotation()),
+        for i in 0.. {
+            match self.child(i) {
+                Child::Empty => generic.push_empty(),
+                Child::Leaf(leaf) => generic.push_leaf(leaf),
+                Child::Node(link) => generic.push_link(link),
+                Child::EndOfNode => break,
+            }
         }
+
+        generic
     }
+
+    /// Construct a specific compound tree from a generic tree
+    fn from_generic(tree: &GenericTree) -> Result<Self, CanonError>
+    where
+        Self::Leaf: Canon,
+        A: Canon;
 }
 
-pub struct ChildIterator<'a, C, A> {
+/// An iterator over the sub-annotations of a Compound collection
+pub struct AnnoIter<'a, C, A> {
     node: &'a C,
     ofs: usize,
     _marker: PhantomData<A>,
 }
 
-impl<'a, C, A> Iterator for ChildIterator<'a, C, A>
+impl<'a, C, A> Clone for AnnoIter<'a, C, A> {
+    fn clone(&self) -> Self {
+        AnnoIter {
+            node: self.node,
+            ofs: self.ofs,
+            _marker: self._marker,
+        }
+    }
+}
+
+impl<'a, C, A> Iterator for AnnoIter<'a, C, A>
 where
     C: Compound<A>,
-    C::Leaf: 'a,
     A: Annotation<C::Leaf> + 'a,
 {
-    type Item = IterChild<'a, C, A>;
+    type Item = WrappedAnnotation<'a, C, A>;
 
     fn next(&mut self) -> Option<Self::Item> {
         loop {
@@ -110,11 +126,11 @@ where
                 Child::EndOfNode => return None,
                 Child::Leaf(l) => {
                     self.ofs += 1;
-                    return Some(IterChild::Leaf(l));
+                    return Some(WrappedAnnotation::Owned(A::from_leaf(l)));
                 }
                 Child::Node(a) => {
                     self.ofs += 1;
-                    return Some(IterChild::Node(a));
+                    return Some(WrappedAnnotation::Link(a.annotation()));
                 }
             }
         }